usb: fix incorrect USB split scheduling
If ep->hcpriv is not set to NULL after completion, it would be reused in a next
transfer. Depending on how long it's been since the previous transmit, it could
be that that a new transmit is delayed for more 1.024 seconds.

This is because the sched_frame will be equal to the sched_frame of the
previous SSPLIT + 10. However, if the previous transmit was 1025 ms ago, the
microframecounter would currently be 0x3FFF / 2 further, resulting in the
dwc_frame_num_gt() check in assign_and_init_hc to fail for 0x3FFF / 2
microframes, which corresponds to 0x1FFF * 125us = 1.024 seconds.

By setting the ep->hcpriv to NULL, a new, correct, scheduling time will be
calculated on each URB transmit request.

Signed-off-by: Sam Van Den Berge <sam.van.den.berge@telenet.be>